Added support for wikidata json format
Using language 'wikidata' in the existing code will download the older
wikidata xml format, because its url follows the convention of the
per-language wikis, e.g. enwiki vs. wikidatawiki.  But
https://www.wikidata.org/wiki/Wikidata:Database_download recommends
using the json dump instead.  It lives in a subdir.  This change is a
ugly but small hack to make the pseudo-language 'entities' download this
file.  It's awkward since the rest of the code really want to be
per-language while wikidata is not language-specific.